+++ NEWS.md
@@ -1,3 +1,6 @@
+## gDRcore 1.3.2 - 2024-06-04
+* switch to `get_supported_experiments`
+
 ## gDRcore 1.3.1 - 2024-05-27
 * synchronize Bioconductor and GitHub versioning